Module type Content_addressable.S

A content-addressable store is an indexed read-write store in which values are keyed directly by their hashes.

Read-only stores

Read-only stores are store where it is only possible to read existing values.

type -'a t

The type for stores. The 'a phantom type carries information about the store mutability.

type key = G.Hash.t

The type for keys.

type value

The type for raw values.

val mem : [> Irmin.Perms.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Irmin.Perms.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val add : [> Irmin.Perms.write ] t -> value -> key Lwt.t

Write the contents of a value to the store. It's the responsibility of the content-addressable store to generate a consistent key.

val unsafe_add : [> Irmin.Perms.write ] t -> key -> value -> unit Lwt.t

Same as add but allows specifying the key directly. The backend might choose to discard that key and/or can be corrupt if the key scheme is not consistent.

val close : 'a t -> unit Lwt.t

close t frees up all the resources associated with t. Any operations run on a closed handle will raise Closed.

val batch : Irmin.Perms.read t -> ([ Irmin.Perms.read | Irmin.Perms.write ] t -> 'a Lwt.t) -> 'a Lwt.t

batch t f applies the writes in f in a separate batch. The exact guarantees depend on the implementation.