Module Tree.Proof

type t

val t : t Irmin.Type.t

val of_tree : tree -> t

of_tree t is the proof representing the tree t. Shallow hashes will be blinded.

val to_tree : t -> tree

of_proof p is the tree representing the proof p. Blinded parts of the proof will raise Dangling_hash when traversed.

val of_keys : tree -> key list -> t Lwt.t

of_keys t keys is the minimal proof that can be used to prove that operations over the domain keys are valid with t.