Module Irmin_git__Commit.Store

Parameters

G : Git.S

Signature

include Irmin.Content_addressable.S with type _ t = bool Stdlib.ref * G.t and type key = G.Hash.t and type value = G.Value.Commit.t

type -'a t

The type for stores. The 'a phantom type carries information about the store mutability.

type key

The type for keys.

type value

The type for raw values.

val mem : [> Irmin__.Import.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Irmin__.Import.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

type 'a t

val close : 'a t -> unit Lwt.t

close t frees up all the resources associated with t. Any operations run on a closed handle will raise Closed.

val add : [> Irmin__.Import.write ] t -> value -> key Lwt.t

Write the contents of a value to the store. It's the responsibility of the content-addressable store to generate a consistent key.

val unsafe_add : [> Irmin__.Import.write ] t -> key -> value -> unit Lwt.t

Same as add but allows specifying the key directly. The backend might choose to discard that key and/or can be corrupt if the key scheme is not consistent.

type 'a t

val clear : 'a t -> unit Lwt.t

Clear the store. This operation is expected to be slow.

type 'a t

val close : 'a t -> unit Lwt.t

close t frees up all the resources associated with t. Any operations run on a closed handle will raise Closed.

type 'a t

val batch : Irmin__.Import.read t -> ([ Irmin__.Import.read | Irmin__.Import.write ] t -> 'a Lwt.t) -> 'a Lwt.t

batch t f applies the writes in f in a separate batch. The exact guarantees depend on the implementation.

module Info = Irmin.Info.Default

module Hash : Irmin.Hash.S with type t = key

module Val : Irmin.Commit.S with type t = value and type hash = key and module Info = Info