Module Irmin_git.Content_addressable

Parameters

module G : Git.S

module V : Irmin.Type.S

Signature

type 'a t = bool Stdlib.ref * G.t

The type for content-addressable backend stores. The 'a phantom type carries information about the store mutability.

type key = G.Hash.t

The type for keys.

type value = V.t

The type for raw values.

val mem : [> `Read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> `Read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val add : [> `Write ] t -> value -> key Lwt.t

Write the contents of a value to the store. It's the responsibility of the content-addressable store to generate a consistent key.

val unsafe_add : [> `Write ] t -> key -> value -> unit Lwt.t

Same as add but allows to specify the key directly. The backend might choose to discared that key and/or can be corrupt if the key scheme is not consistent.