Module Irmin_git.Content_addressable

Use Git as a content-addressable store. Values will be stored into .git/objects.

Parameters

module G : Git.S

Signature

module type S = Irmin.Content_addressable.S with type _ t = G.t and type key = G.Hash.t

A content-addressable store is an indexed read-write store in which values are keyed directly by their hashes.

module Make (V : Irmin.Type.S) : S with type value = V.t

A content-addressable store is an indexed read-write store in which values are keyed directly by their hashes.